requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.25.0
celery==5.3.6
redis==5.0.1
sqlalchemy==2.0.28
asyncpg==0.29.0
python-dotenv==1.0.1
httpx==0.25.2
tenacity==8.2.3
pydantic-settings==2.2.1
psycopg2-binary==2.9.9
python-telegram-bot==20.6
openai 